complex plot in second level function, plotting
