+++ examples/proc_square_refl.py
@@ -1,59 +1,10 @@
 from pylab import *
 from pyspecdata import *
+from proc_scripts import *
 from proc_scripts.third_level.analyze_square_refl import analyze_square_refl
 init_logging("debug")
 
-class fl_ext(figlist_var):
-    def next(self, *arg, **kwargs):
-        kwargs.update({"figsize": (9, 6), "legend": True})
-        super().next(*arg, **kwargs)
-
-    def complex_plot(fl, d, label="", show_phase=False, show_real=True):
-        colors = []
-        for j in range(ndshape(d)["ch"]):
-            chlabel = d.getaxis("ch")[j]
-            l = fl.plot(
-                abs(d["ch", j]),
-                linewidth=3,
-                alpha=0.5,
-                label="CH%d abs " % chlabel + label,
-            )
-            colors.append(l[0].get_color())
-            if show_real:
-                fl.plot(
-                    d["ch", j].real,
-                    linewidth=1,
-                    color=colors[-1],
-                    alpha=0.5,
-                    label="CH%d real " % chlabel + label,
-                )
-                fl.plot(
-                    d["ch", j].imag,
-                    "--",
-                    linewidth=1,
-                    color=colors[-1],
-                    alpha=0.5,
-                    label="CH%d imag " % chlabel + label,
-                )
-            fl.grid()
-            if show_phase:
-                fl.twinx(orig=False)
-                fl.plot(
-                    d["ch", j].angle/2/pi,
-                    ".",
-                    linewidth=1,
-                    color=colors[-1],
-                    alpha=0.3,
-                    label="CH%d angle " % chlabel + label,
-                )
-                ylabel("phase / cyc", size=10)
-                ax2=gca()
-                gridandtick(ax2, use_grid=False)
-                ax2.grid(False)
-                fl.twinx(orig=True)
-        return colors
-
-with fl_ext() as fl:
+with fl_mod() as fl:
     for filename, expno, dataset_name in [("201218_sqwv_cap_probe_2", "capture1", "hairpin probe"),
             ('201228_sqwv_sol_probe_1', 'capture1', 'solenoid probe')]:
         print("processing dataset",dataset_name)
